feat : 添加裁判系统接口

+++ src/sentry_behavior_tree/plugins/condition/low_hp_condition.cpp
@@ -6,7 +6,7 @@ namespace sentry_behavior_tree
 
     BT::NodeStatus lowHp(BT::TreeNode & tree_node){
 
-        auto base_unfolds = tree_node.config().blackboard->get<int>("robot_hp") <= 150;
+        auto base_unfolds = tree_node.config().blackboard->get<u_int16_t>("robot_hp") <= 150;
 
         return base_unfolds ? BT::NodeStatus::SUCCESS : BT::NodeStatus::FAILURE;
